Fixed folder spelling mistake
